backend/PubSub/requirements.txt
functions-framework==3.*
google-cloud-pubsub
requests
google-cloud-firestore
firebase-admin==6.0.0